+++ R/guide-axis-theta.R
@@ -41,6 +41,7 @@ guide_axis_theta <- function(title = waiver(), theme = NULL, angle = waiver(),
     angle = angle,
     cap = cap,
     minor.ticks  = minor.ticks,
+    theme = theme,
 
     # parameter
     available_aes = c("x", "y", "theta"),
@@ -85,8 +86,6 @@ GuideAxisTheta <- ggproto(
     key <- params$key
     n <- nrow(key)
 
-    params$theme_aes <- coord$theta %||% params$aesthetic
-
     if (!("theta" %in% names(key))) {
       # We likely have a linear coord, so we match the text angles to
       # standard axes to be visually similar.
@@ -129,12 +128,13 @@ GuideAxisTheta <- ggproto(
   },
 
   setup_elements = function(params, elements, theme) {
+    theme <- add_theme(theme, params$theme)
 
     axis_elem <- c("line", "text", "ticks", "minor", "major_length", "minor_length")
     is_char <- vapply(elements[axis_elem], is.character, logical(1))
     axis_elem <- axis_elem[is_char]
 
-    aes <- switch(
+    aes <- params$theme_suffix %||% switch(
       params$position,
       theta     = "x.bottom",
       theta.sec = "x.top",
@@ -306,7 +306,7 @@ GuideAxisTheta <- ggproto(
     list(offset = max(height))
   },
 
-  arrange_layout = function(key, sizes, params) {
+  arrange_layout = function(key, sizes, params, elements) {
     NULL
   },
 
@@ -327,7 +327,7 @@ GuideAxisTheta <- ggproto(
     if (params$position %in% c("top", "bottom")) {
       height <- sum(
         elements$offset,
-        unit(max(height_cm(grobs$labels$children)), "cm")
+        unit(max(height_cm(grobs$labels)), "cm")
       )
       vp <- viewport(
         y = unit(as.numeric(params$position == "bottom"), "npc"),
@@ -337,7 +337,7 @@ GuideAxisTheta <- ggproto(
     } else {
       width <- sum(
         elements$offset,
-        unit(max(width_cm(grobs$labels$children)), "cm")
+        unit(max(width_cm(grobs$labels)), "cm")
       )
       vp <- viewport(
         x = unit(as.numeric(params$position == "left"), "npc"),